refactor: add link, change style
