+++ src/app/components/EventBlock/EventBlock.scss
@@ -1,0 +1,8 @@
+.event-link {
+  text-decoration: none;
+
+  .event__name {
+    font-weight: 600;
+    text-decoration: underline;
+  }
+}
