+++ src/app/pages/News/DetailNews/Interview_EuropeSuffer.html
@@ -1,0 +1,23 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <link rel="icon" type="image/svg+xml" href="/favicon.png" />
+  <meta name="description" content="Серии ранних рисунков «Базис» и «Пылесосы в пейзаже» художницы Катерины Ковалёвой.">
+
+
+  <title data-localize="artworks">Интервью Европа страдает Парашют который нас спасёт</title>
+</head>
+
+<body>
+  <main>
+    <p>Интервью Европа страдает Парашют который нас спасёт</p>
+  </main>
+  <script type="module" src="/src/index.ts"></script>
+</body>
+
+</html>
