+++ src/app/pages/News/news.scss
@@ -20,9 +20,7 @@
   cursor: pointer;
 }
 
-.pagination__page-number {
-  // margin: 0 5px;
-}
+.pagination__page-number {}
 
 .pagination__page-number--active {
   text-decoration: underline;
