+++ src/style/announcement.scss
@@ -50,15 +50,7 @@ div {
 
 
 .event {
-
-
-  // box-sizing: border-box;
   display: block !important;
-  // flex-direction: column;
-  // justify-content: center;
-  // align-items: center;
-
-  // gap: 20px;
 
   padding: 12px 12px 12px 12px;
   max-width: 384px;
@@ -103,7 +95,6 @@ div {
     margin-bottom: 15px;
 
     font-size: 18px;
-    // text-transform: uppercase;
     text-align: left;
   }
