+++ src/style/main.scss
@@ -29,3 +29,4 @@
 @import '/src/app/pages/FancyBox/fancybox.scss';
 
 @import '/src/app/pages/News/news.scss';
+@import '/src/app/components/EventBlock/EventBlock.scss';
